Update JDK version and refactor test packages/models
Upgraded JDK from version 23 to 24 in the GitHub Actions workflow. Refactored test package structure for better clarity and updated OpenAI model references from "gpt-4o-mini" to "gpt-4.1-nano". Minor code optimizations and enhancements were also made across various tests.

Update Kotlin version to 2.1.21
