+++ gradle.properties
@@ -4,4 +4,4 @@ org.gradle.jvmargs=-Xms512M -Xmx1G -Dfile.encoding=UTF-8 -Dconsole.encoding=UTF-
 org.gradle.configureondemand=true
 #Kotlin
 kotlin.code.style=official
-defaultKotlinVersion=2.1.20
+defaultKotlinVersion=2.1.21
