+++ src/test/kotlin/e01/JavaAsyncWrapperTest.kt
@@ -3,12 +3,14 @@ package e01
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
+import kotlin.coroutines.CoroutineContext
 import kotlin.test.Test
 import kotlin.time.Duration.Companion.seconds
 
 class JavaAsyncWrapperTest {
-    suspend fun chatAsync(): String {
-        return withContext(Dispatchers.IO) {
+
+    suspend fun chatAsync(context: CoroutineContext = Dispatchers.IO): String {
+        return withContext(context) {
             Thread.sleep(1.seconds.inWholeMilliseconds)
             return@withContext "Hello, World"
         }
